consolidate internal testing
